Pulsar/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
black==22.12.0
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
docformatter==1.5.0
flake8==6.0.0
isort==5.11.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
unify==0.5
untokenize==0.1.1
